Fix multi branch recursion
Prior to this trivial fix, multiple recursive calls in the same
expression would fail due to the first recursive call messing up the
scope for the remaining recursive calls. This happened as only ONE
(new)scope was used every time a block was run. This was the scope that
was setup for the block when the block was defined. In other words,
previously, a scope for a block was created only once i.e., when the
block was defined and it was reused on every invocation of the block.
All operations on the scope (setting, etc) would thus show up in the
subsequent executions of the block. This was an incorrect way of
operating and thinking about function execution.

In moks, all blocks create a new scope. Thus, when the are run, they
should always create a new scope and run within that scope. This model
makes sense and gives us a benefit of supporting recursion by default.

Also, thanks to @jamak for helping me figure out how multi branch
recursion works and to help me understand this issue conceptually.