Add clone dependency
Since require caches imports, when we use an already imported file
in createScope statement in loadMoksModule, we end up getting back
the same file which is not what we need as we need a brand new
copy where we can set the __parent__ property to what we want.
We need a fresh import basically and not stuff thats already in
use and thus has stuff we have done to it (like add __parent__)
already on it.